+++ src/test/scala/ParserTest.scala
@@ -6,6 +6,7 @@ import server.messages.query.DatabaseMessages._
 import server.messages.query.MapMessages._
 import server.messages.query.RowMessages._
 import ErrorMessages.InvalidQueryMessage
+import server.messages.query.LoginMessage
 import server.util.Parser
 
 /**
